fix(Net):  load error

+++ GanHuo/Extension/UIColor+Extension.swift
@@ -67,7 +67,7 @@ extension UIColor {
         self.init(red: (CGFloat((rgbValue & 0xFF0000) >> 16)) / 255.0, green: (CGFloat((rgbValue & 0xFF00) >> 8)) / 255.0, blue: (CGFloat(rgbValue & 0xFF)) / 255.0, alpha: 1.0)
     }
     
-    convenience init(rgbaValue: Int) {
+    convenience init(rgbaValue: Int64) {
         self.init(red: (CGFloat((rgbaValue & 0xFF000000) >> 24)) / 255.0, green: (CGFloat((rgbaValue & 0xFF0000) >> 16)) / 255.0, blue: (CGFloat((rgbaValue & 0xFF00) >> 8)) / 255.0, alpha: (CGFloat(rgbaValue & 0xFF)) / 255.0)
     }
